Move seed data out of the published API
Bogus and SeedTestData shipped in the API image although they are only
used for local development. They now live in Geopilot.Api.SeedData,
which references the API and never the other way round, so publishing
Geopilot.Api cannot reach them. A reference back would be circular and
fail the build, and DependencyBoundaryTest fails if the API depends on
either of them again.

The seed no longer runs inside the API process. A "seed" compose service
runs it instead, which covers all three development flows: the
"Development" launch profile, a single startup project, and a plain
docker compose up. Its Dockerfile stages sit above "final" so the
default build target stays the API image.

Because the seed can now run after someone signed in, it only fills a
database that has neither users nor mandates. In development the API
logs where the test data comes from when no mandates exist, so an empty
database is not mistaken for a broken setup.
